swagger-version

+++ apps/distribute-service/src/controllers/post.controller.ts
@@ -1,10 +1,11 @@
-import { Controller, Post, Get, Query, Body, UseGuards, Patch, Delete, Param} from "@nestjs/common";
+import { Controller, Post, Get, Query, Body, UseGuards, Patch, Delete, Param, Req, UnauthorizedException} from "@nestjs/common";
 import {MainPostService} from '../providers/post.service'
 import { JWTAuthGuard, ZodValidationPipe } from "@repo/api";
 import {CreationPostDataSchema, type CreationPostDataType, UpdatetingPostDataSchema,
      type UpdatePostDataType, DeletePostDataSchema, type DeletePostDataType} from '@repo/user-interfaces'
 import { ApiBody, ApiCreatedResponse, ApiNoContentResponse, ApiNotAcceptableResponse, ApiOkResponse, ApiOperation, ApiQuery, ApiResponse, ApiTags, ApiBearerAuth, ApiCookieAuth } from "@nestjs/swagger";
 import { CretionNewPostSwagger, DeletePostSwagger, Post_Entity_Swagger, UpdationPostSwagger } from "src/documentation_classes/post.swagger";
+import { type Request } from "express";
 
 @ApiTags('post')
 @Controller('post')
@@ -38,9 +39,12 @@ export class MainPostCOntriller {
     @ApiBody({type: CretionNewPostSwagger})
     @ApiResponse({status: 201, description: 'creation new post successfully'})
     async handleCreateNewPost(
+        @Req() req: Request,
         @Body(new ZodValidationPipe(CreationPostDataSchema)) dto: CreationPostDataType
     ) : Promise<void> {
-        await this.postService.creationNewPost(dto)
+        if (!req.user) throw new UnauthorizedException('you dont have the token')
+            
+        await this.postService.creationNewPost({...dto, userId: req.user.userId})
     }
 
     @ApiCookieAuth()
@@ -51,10 +55,16 @@ export class MainPostCOntriller {
     @ApiBody({type: UpdationPostSwagger})
     @ApiNotAcceptableResponse({description: 'you update your post succesfully'})
     async handleUpdatePost(
+        @Req() req: Request,
         @Param("id") id: string,
         @Body(new ZodValidationPipe(UpdatetingPostDataSchema)) dto: UpdatePostDataType
     ): Promise<void> {
-        await this.postService.updatePost({...dto, id: Number(id)})
+        if (!req.user) throw new UnauthorizedException('you dont have token (post)')
+
+        console.log(dto);
+        
+
+        await this.postService.updatePost({...dto, id: Number(id), userId: req.user.userId})
     }
 
     @ApiCookieAuth()
@@ -65,8 +75,10 @@ export class MainPostCOntriller {
     @ApiOperation({summary: 'delete', description: 'deleted post'})
     @ApiNoContentResponse({description: 'your post delete successfully'})
     async handleDeletePost(
+        @Req() req: Request,
         @Param('id') id: string
     ) : Promise<void> {
-        await this.postService.deletePost({id: Number(id)})
+        if (!req.user) throw new UnauthorizedException('you dont have the token')
+        await this.postService.deletePost({id: Number(id), userId: req.user.userId})
     }
 }
